修改10/02-03-04错误处理

+++ 10-file-system/04-my-rm.js
@@ -1,4 +1,9 @@
 #!/usr/bin/node
 const fs = require('fs'),
       file = process.argv[2];
-fs.unlinkSync(file);
+if(fs.existsSync(file)){
+  fs.unlinkSync(file);
+}else{
+  console.error('%s not exist!',file);
+  process.exit(1);
+}
